simple confirmatoin before delete user/lyrics

+++ src/app/admin/dashboard/dashboard.component.ts
@@ -7,6 +7,8 @@ import {Router} from "@angular/router";
 import {ThemeService} from "../../shared/services/settings/theme.service";
 import {UsersService} from "../users.service";
 import {User} from "../../interfaces/user";
+import {MatDialog} from "@angular/material/dialog";
+import {ConfirmDialogComponent} from "../shared/components/confirm-dialog/confirm-dialog.component";
 
 @Component({
   selector: 'app-dashboard',
@@ -22,12 +24,15 @@ export class DashboardComponent implements OnInit {
   deleteUserSub: Subscription | undefined;
   theme: string | undefined;
   public endSession: Date | undefined;
+  title: string = '';
+  message: string = '';
 
   constructor(private lyricsService: LyricsService,
               private auth: AuthService,
               private router: Router,
               private themeService: ThemeService,
-              private usersService: UsersService) { }
+              private usersService: UsersService,
+              public dialog: MatDialog) { }
 
   ngOnInit(): void {
     this.themeService.getCurrentTheme().subscribe( theme => this.theme = theme);
@@ -54,16 +59,20 @@ export class DashboardComponent implements OnInit {
   }
 
 
-  removeLyrics(id: string | any) {
-    this.deleteLyricsSub = this.lyricsService.removeLyrics(id).subscribe( () => {
-      this.lyricsList = this.lyricsList.filter(list => list.id !== id);
-    })
+  removeLyrics(id: string | any, name: string) {
+    if(confirm("Are you sure to delete " + name)){
+      this.deleteLyricsSub = this.lyricsService.removeLyrics(id).subscribe( () => {
+        this.lyricsList = this.lyricsList.filter(list => list.id !== id);
+      })
+    }
   }
 
-  removeUser(id: string | any) {
-    this.deleteUserSub = this.usersService.removeUser(id).subscribe( () => {
-      this.users = this.users.filter(list => list.id !== id);
-    })
+  removeUser(id: string | any, email: string) {
+    if(confirm("Are you sure to delete " + email)) {
+      this.deleteUserSub = this.usersService.removeUser(id).subscribe(() => {
+        this.users = this.users.filter(list => list.id !== id);
+      })
+    }
   }
 
   getAuthTokenExp(){
@@ -82,4 +91,17 @@ export class DashboardComponent implements OnInit {
     // @ts-ignore
     return this.endSession.toLocaleString();
   }
+
+  openDialog(): void{
+    const dialogRef = this.dialog.open(ConfirmDialogComponent, {
+      width: '250px',
+      data: {title: this.title, message: this.message},
+    });
+
+    dialogRef.afterClosed().subscribe(result => {
+      this.message = result;
+    })
+
+  }
+
 }
